refactor(finish): move the finish RunOperations to src/operations/
`src/finish/operations/` was the only place in the repo holding RunOperations
outside `src/operations/`. project-conventions.md is explicit: "New Operations:
src/operations/<name>.ts, exported from the barrel." The design doc that drove
the port sketched them under src/finish/, and that sketch was followed over the
repo's own convention.

  review-op.ts    -> src/operations/finish-review.ts
  fix-op.ts       -> src/operations/finish-fix.ts
  narrative-op.ts -> src/operations/finish-narrative.ts

The flat `finish-*` naming matches the siblings already there (acceptance-fix,
plan-draft, debate-judge), and the ops are exported from the operations barrel
like every other op. Their tests move with them, per the test/unit mirror rule.

pr-title.ts does NOT move: it is pure title helpers (sanitize/parse/resolve and
their constants), not an operation, and it is shared — src/finish/pr/context.ts
uses resolveTitle. It moves up one level to src/finish/pr-title.ts, out of the
deleted directory, and stays exported from @/finish.

No import cycle is introduced, which is why the ops import finish's leaf modules
(../finish/review, ../finish/types, ../finish/ops, ../finish/pr-title) rather
than the @/finish barrel: the barrel reaches @/cli, and pulling that into
@/operations — which nearly everything imports — would be a load-time chain.
None of those leaf modules import @/operations, so the edge stays one-way. This
matches how the existing ops already reach into feature modules (../review/*,
../acceptance/types, ../tdd/*).

Pure move: no behaviour change, no logic edited.
