+++ test/unit/operations/finish-fix.test.ts
@@ -9,8 +9,8 @@ import { join } from "node:path";
 import { afterEach, describe, expect, test } from "bun:test";
 import type { ConfigSelector } from "@/config";
 import type { FinishConfig } from "@/config/selectors";
-import type { FinishFixInput } from "@/finish";
-import { finishFixOp } from "@/finish";
+import type { FinishFixInput } from "@/operations";
+import { finishFixOp } from "@/operations";
 import type { Finding } from "@/finish";
 import type { NaxRuntime } from "@/runtime";
 import { makeTestRuntime, withTempDir } from "@test/helpers";
